Initial Game Entity code
Added GameEntity class and it's manager.
Added code to bind to Lua
GameEntityManager is a singleton.
Worked out how to bind the singleton getSingleton member function to Lua.  No need for a 'thunk'.
GameEntity is designed as a C++ base class for Lua classes.
GameEntity::update can be (and should be) overiden in Lua.

Some minor updates to scripts.  Mainly to have something intresting to look at while debuging. :-)