Add keyboard handling to gui controller.
Also changed to call mouseMoved and keypressed as object methods, so
that the top level Panel can be registered with the controller rahter
than the overlay.  This saves having to write mouse and keyboard handers
for each overlay instance.

Added keyboard handler to Panel and Button.  Checkbox inherits this from
button.  A key can be set on buttons, to cause the same action as
clicking.  Checkboxes also inherit this behaviour.