First commit of new mouse picking code.
Giving a 2D x,y position, a ray is fired from the POV of the camera.  Game entities with their meshes
bounding box intersected are returned and with the Lua version the distance is returned also.

GameEntityManager now has initialise and shutdown methods.

TODO

Method to return the position the mesh (not BB) was intersected, as a Vector3.  This will be used for
directing navigation, and checking that a mesh was actually intersected, and not just its bounding box.