A star implementaion.
Functions to path find through the navigation mesh.  Some ideas from Game Programming Gems 1 - "A* Speed Optimizations",
namely using std::push_heap etc to implement a priority queue, while retainting the ability to "promote" a search node, when
its costs change for the better.  Only an open list is maintained, the closed list is represented as a state on each node (cell).

Implementation is simlified somewhat by holding node information in the cells, and not doing incremental searches.  This
will only be fast enough with a small data set.