Navigation Path Straightening and Path test code.
Code to take a generated path, and straighten it, by removing turns that
are unnecessary to remain within the navigation mesh.  A maximum turn
angle is enforced, as well as pinning any points where the path starts
going up or down from its last direction.

Temporary code to test all this added to engine.cpp