Ward pointers and LuaBridge
Lua Bridge slightly modfied to always call a containers 'get' pointer for every access rather than just on instantation.

This allows ward pointer to do its checking.

Ward pointer changed to save the co-responding data in a map so when a new ward pointer is created for the same raw pointer, then previous ref count and valid flag is maintained.